backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
python-dotenv>=1.0.0
sentence-transformers>=3.4.0
umap-learn>=0.5.6
faiss-cpu>=1.10.0
numpy>=1.26.0
scipy>=1.13.0
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0
tree-sitter-typescript==0.21.0
tree-sitter-java==0.21.0
tree-sitter-go==0.21.0
ollama>=0.3.0
diskcache>=5.6.0
